build, mount: Fix 32-bit build for mfsmount3
There was a problem with compiling mfsmount3 on 32-bit systems because
FUSE3 doesn't use uint64_t on 32-bit systems (it uses unsigned long
instead). This commit fixes the problem.

+++ src/chunkserver/iostat.h
@@ -78,7 +78,7 @@ public:
 			if (ret < 0) {
 				continue;
 			}
-			stats_[st.st_dev] = {0, stfs.f_blocks};
+			stats_[st.st_dev] = {0, (size_t)stfs.f_blocks};
 		}
 	}
